+++ brainbox/behavior/wheel.py
@@ -4,7 +4,7 @@ Set of functions to handle wheel data
 import numpy as np
 from numpy import pi
 import scipy.interpolate as interpolate
-from scipy.signal import convolve, windows
+import scipy.signal
 from scipy.linalg import hankel
 import matplotlib.pyplot as plt
 from matplotlib.collections import LineCollection
@@ -92,6 +92,21 @@ def velocity(re_ts, re_pos):
         return ifcn(re_ts)
 
 
+def velocity_filtered(pos, fs, corner_frequency=20, order=8):
+    """
+    Compute wheel velocity from uniformly sampled wheel data
+
+    :param pos: vector of uniformly sampled wheel positions
+    :param fs: scalar, sampling frequency
+    :param corner_frequency: scalar, corner frequency of low-pass filter
+    :param order: scalar, order of Butterworth filter
+    """
+    sos = scipy.signal.butter(**{'N': order, 'Wn': corner_frequency / fs * 2, 'btype': 'lowpass'}, output='sos')
+    vel = np.insert(np.diff(scipy.signal.sosfiltfilt(sos, pos)), 0, 0) * fs
+    acc = np.insert(np.diff(vel), 0, 0) * fs
+    return vel, acc
+
+
 def velocity_smoothed(pos, freq, smooth_size=0.03):
     """
     Compute wheel velocity from uniformly sampled wheel data
@@ -116,12 +131,12 @@ def velocity_smoothed(pos, freq, smooth_size=0.03):
     std_samps = np.round(smooth_size * freq)  # Standard deviation relative to sampling frequency
     N = std_samps * 6  # Number of points in the Gaussian covering +/-3 standard deviations
     gauss_std = (N - 1) / 6
-    win = windows.gaussian(N, gauss_std)
+    win = scipy.signal.windows.gaussian(N, gauss_std)
     win = win / win.sum()  # Normalize amplitude
 
     # Convolve and multiply by sampling frequency to restore original units
-    vel = np.insert(convolve(np.diff(pos), win, mode='same'), 0, 0) * freq
-    acc = np.insert(convolve(np.diff(vel), win, mode='same'), 0, 0) * freq
+    vel = np.insert(scipy.signal.convolve(np.diff(pos), win, mode='same'), 0, 0) * freq
+    acc = np.insert(scipy.signal.convolve(np.diff(vel), win, mode='same'), 0, 0) * freq
 
     return vel, acc
 
@@ -274,8 +289,8 @@ def movements(t, pos, freq=1000, pos_thresh=8, t_thresh=.2, min_gap=.1, pos_thre
     peak_amps = np.fromiter(peaks, dtype=float, count=onsets.size)
     N = 10  # Number of points in the Gaussian
     STDEV = 1.8  # Equivalent to a width factor (alpha value) of 2.5
-    gauss = windows.gaussian(N, STDEV)  # A 10-point Gaussian window of a given s.d.
-    vel = convolve(np.diff(np.insert(pos, 0, 0)), gauss, mode='same')
+    gauss = scipy.signal.windows.gaussian(N, STDEV)  # A 10-point Gaussian window of a given s.d.
+    vel = scipy.signal.convolve(np.diff(np.insert(pos, 0, 0)), gauss, mode='same')
     # For each movement period, find the timestamp where the absolute velocity was greatest
     peaks = (t[m + np.abs(vel[m:n]).argmax()] for m, n in zip(onset_samps, offset_samps))
     peak_vel_times = np.fromiter(peaks, dtype=float, count=onsets.size)